refactor: Active Filter Display
